ENH: Adds batch normalization for preventing layer output explosion
Batch normalization is added to the Encoder until the Dense layers that output z_mean and z_log_var, so the Reparametrization Trick does not explode when computing exp(0.5*z_log_var). Added also a lower variance for sampling epsilon in the Reparametrization to reduce the stochasticity of the process, which should improve the reconstruction, but hurt the generalizability a bit.