BUGFIX: condVAE model loading problem when using non-default architecture parameters
The wrapping class 'IncrFeatStrided...' was not instantiating any config, the default architecture parameters were being used for loading the model, so when using (e.g.) latent_space_dims=64 when training, when wanting to load, the loader was thinking that latent_space_dims=32 because no config had been specified in the wrapping class. Added config functions to the class, now models with non-default initialization architectural parameters can be loaded.